% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Resource Rolodex Data}
\format{
A data frame with rows and columns
\describe{
  \item{Name}{Resource Name}
  \item{Address}{Resource Address}
  \item{Phone}{Resource Phone number}
  \item{Website}{Resource website}
  \item{Hours}{Resource hours of operation}
  \item{Email address}{Resource contact email}
  \item{Type}{Resource category}
  \item{Loc}{Resource location on or off campus}
  \item{Info}{Resource additional description}
}
}
\source{
generated by Project ACCESS Navigators
}
\usage{
data
}
\description{
On- and Off-Campus resources and services for UAlbany Students
}
\keyword{datasets}
